chore(core): move types from types dir to current files, remove types dir and adapt package.json
