--- packages/botonic-core/index.d.ts
@@ -1,2 +1,0 @@
-// eslint-disable-next-line node/no-unpublished-import
-export * from './src'
