+++ packages/botonic-core/package.json
@@ -4,7 +4,6 @@
   "license": "MIT",
   "description": "Build Chatbots using React",
   "main": "lib/index.js",
-  "types": "./src/types",
   "scripts": {
     "test": "../../node_modules/.bin/jest --coverage",
     "prepare": "node ../../preinstall.js",
