+++ packages/botonic-core/src/constants.ts
@@ -1,16 +1,15 @@
-// @ts-nocheck
 export const Providers = Object.freeze({
   Messaging: {
-    TELEGRAM: 'telegram',
-    WECHAT: 'wechat',
     FACEBOOK: 'facebook',
-    INTERCOM: 'intercom',
     GENERIC: 'generic',
-    TWITTER: 'twitter',
-    WHATSAPP: 'whatsapp',
-    SMOOCH: 'smooch',
-    SMOOCH_WEB: 'smooch_web',
     IMBEE: 'imbee',
+    INTERCOM: 'intercom',
+    SMOOCH_WEB: 'smooch_web',
+    SMOOCH: 'smooch',
+    TELEGRAM: 'telegram',
+    TWITTER: 'twitter',
     WEBCHAT: 'webchat',
+    WECHAT: 'wechat',
+    WHATSAPP: 'whatsapp',
   },
 })
