+++ packages/botonic-core/src/errors.ts
@@ -1,6 +1,8 @@
-// @ts-nocheck
+import { Input } from './index'
+
 export class NoMatchingRouteError extends Error {
-  constructor(input) {
+  input: Input
+  constructor(input: Input) {
     super(
       `No route found for input '${String(input)}' and no 404 route defined`
     )
