+++ packages/botonic-core/src/i18n.ts
@@ -1,6 +1,11 @@
-// @ts-nocheck
 // eslint-disable-next-line filenames/match-regex
-export const getString = (locales, locale, stringID) => {
+import { Locales } from './index'
+
+export const getString = (
+  locales: Locales,
+  locale: string,
+  stringID: string
+): string => {
   let l = stringID.split('.').reduce((o, i) => o[i], locales[locale])
   if (l instanceof Array) l = l[Math.floor(Math.random() * l.length)]
   return String(l)
