--- packages/botonic-core/src/types/constants.d.ts
@@ -1,14 +1,0 @@
-export const Providers: Readonly<{
-  Messaging: {
-    FACEBOOK: 'facebook'
-    GENERIC: 'generic'
-    IMBEE: 'imbee'
-    INTERCOM: 'intercom'
-    SMOOCH: 'smooch'
-    TELEGRAM: 'telegram'
-    TWITTER: 'twitter'
-    WEBCHAT: 'webchat'
-    WECHAT: 'wechat'
-    WHATSAPP: 'whatsapp'
-  }
-}>
