--- packages/botonic-core/src/types/debug/index.d.ts
@@ -1,1 +1,0 @@
-export * from './inspector'
