--- packages/botonic-core/src/types/i18n.d.ts
@@ -1,8 +1,0 @@
-/* eslint-disable filenames/match-regex */
-import { Locales } from '.'
-
-export declare function getString(
-  locales: Locales,
-  locale: string,
-  stringID: string
-): string
