--- packages/botonic-core/src/types/plugins.d.ts
@@ -1,13 +1,0 @@
-import { Input, PluginConfig, Session } from '.'
-
-type PluginMode = 'pre' | 'post'
-
-export declare function loadPlugins(plugins: PluginConfig<any>[]): any
-export declare function runPlugins(
-  plugins: PluginConfig<any>[],
-  mode: PluginMode,
-  input: Input,
-  session: Session,
-  lastRoutePath: string,
-  response?: any
-): Promise<void>
