--- packages/botonic-core/src/types/utils.d.ts
@@ -1,12 +1,0 @@
-export declare function getWebpackEnvVar(
-  webpackEnvVar: string | boolean,
-  name: string | boolean,
-  defaultValue: string
-): string
-export declare function isFunction(o: any): boolean
-export declare function isBrowser(): boolean
-export declare function isMobile(mobileBreakpoint?: number): boolean
-export declare function isNode(): boolean
-export declare function params2queryString(params: any): string
-export declare function join(...segments: any[]): string
-export declare function dirname(path: any): string
